feat: ssue 421 expedition api (#503)
* feat(#421): replace Playwright expedition print with Shoptet REST API + QuestPDF

- Add ShoptetApiExpeditionListSource: fetches orders via Shoptet REST API
  (8 shipping methods: Zásilkovna, PPL, GLS; Osobní odběr batch=1)
- Add ShoptetApiExpeditionClient: typed HttpClient for Shoptet order API
  with configurable base URL and bearer token auth
- Add ExpeditionProtocolDocument: QuestPDF code-first picking list PDF
  with barcode (ZXing CODE_128) per order, items table, and summary page
- Add ExpeditionProtocolData / ExpeditionOrder / ExpeditionOrderItem models
- Add ShippingMethod enum with CarrierGroupName and IsOsobak helpers
- Wire up DI in ShoptetApiServiceCollectionExtensions
- Remove ShoptetPlaywrightExpeditionListSource and PrintPickingListScenario
- Add SkiaSharp.NativeAssets.Linux for native barcode rendering on Linux
- Add unit tests: ExpeditionProtocolDocumentTests, ShoptetApiExpeditionListSourceTests

Closes #421

@claude

* feat: refactor

* feat(expedition): improve PDF layout — separate Varianta/Pozice columns, compact sizing

- Add dedicated Varianta and Pozice columns to both order and summary tables
- Remove OK column and inline warehouse/variant text from description cell
- Strip "Obsah: " prefix from variant values at render time
- Reduce margins, font sizes, cell padding, and spacing for more orders per page
- Keep visual inspection test PDF on disk for manual review


* fix: mapping

* fix: missing shippings

* fix: sets

* fix: failing be tests

* ci: proper smoke test (liveness probe only)

* fix: BE tests

* fix: add StockDemand to ExpeditionOrderItem and fix test compilation errors

- Add StockDemand property to ExpeditionOrderItem (maps from Shoptet StockStatus.AllDemand)
- Populate StockDemand in MapOrderItems from item.StockStatus?.AllDemand
- Fix BuildSource in tests to pass required TimeProvider and ICatalogRepository args
- Fix Amount type mismatch (string -> decimal?) in test helper


* feat: add refresh button to ExpeditionListArchivePage with tests

Adds an "Obnovit" button that invalidates all expedition archive queries,
with loading/disabled state during invalidation. Covered by 4 unit tests.


---------

+++ backend/src/Adapters/Anela.Heblo.Adapters.ShoptetApi/Expedition/ExpeditionProtocolData.cs
@@ -22,8 +22,8 @@ public class ExpeditionOrderItem
     public string Variant { get; set; } = null!;
     public string WarehousePosition { get; set; } = null!;
     public int Quantity { get; set; }
-    public int StockCount { get; set; }
-    public int StockDemand { get; set; }
+    public decimal StockCount { get; set; }
+    public decimal StockDemand { get; set; }
     public decimal UnitPrice { get; set; }
     public string Unit { get; set; } = string.Empty;
 }
